backend/requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
attrs==23.1.0
Automat==22.10.0
awscli==1.29.71
beautifulsoup4==4.12.2
billiard==4.2.0
blinker==1.7.0
boto3==1.28.84
botocore==1.31.84
bs4==0.0.1
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.36.0
colorama==0.4.4
constantly==15.1.0
cryptography==41.0.4
cssselect==1.2.0
distro==1.8.0
Django==4.2.7
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==4.3.1
djangorestframework==3.14.0
docker==6.1.3
docutils==0.16
EasyProcess==1.1
entrypoint2==1.1
fake-useragent==1.3.0
Faker==19.13.0
filelock==3.12.4
Flask==3.0.0
Flask-Cors==4.0.0
greenlet==3.0.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.3.4
lxml==4.9.3
MarkupSafe==2.1.3
MouseInfo==0.1.3
mss==9.0.1
numpy==1.26.1
openai==1.2.3
outcome==1.3.0
packaging==23.2
pandas==2.1.1
parsel==1.8.1
playwright==1.39.0
prompt-toolkit==3.0.41
Protego==0.3.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyDispatcher==2.0.7
pyee==11.0.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==10.0
pyobjc-framework-Cocoa==10.0
pyobjc-framework-Quartz==10.0
pyOpenSSL==23.2.0
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
pytweening==1.0.7
pytz==2023.3.post1
PyYAML==6.0.1
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
retrying==1.3.4
rsa==4.7.2
rubicon-objc==0.4.7
s3transfer==0.7.0
scrapinghub==2.4.0
Scrapy==2.11.0
scrapy-fake-useragent==1.4.4
scrapy-zyte-smartproxy==2.3.0
selenium==4.14.0
service-identity==23.1.0
setuptools==68.2.2
shub==2.14.5
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
tldextract==5.0.0
toml==0.10.2
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.11.1
Twisted==22.10.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
w3lib==2.1.2
wcwidth==0.2.12
websocket-client==1.6.4
Werkzeug==3.0.1
wsproto==1.2.0
zope.interface==6.1